Update to handle CKAN rows
