+++ .babelrc
@@ -1,3 +1,20 @@
 {
-  "presets": [ "react", "es2015" ]
+  "presets": [
+    ["env", {
+      "targets": {
+        "browsers": [
+          "last 2 versions",
+          "IE>10"
+        ]
+      },
+      "useBuiltIns": true,
+      "exclude": [
+        "transform-regenerator"
+      ]
+    }],
+    "react"
+  ],
+  "plugins": [
+    "transform-object-rest-spread"
+  ]
 }
