+++ package.json
@@ -13,22 +13,7 @@
   "private": false,
   "repository": "https://github.com/bjacobel/bosfoodfails",
   "dependencies": {
-    "babel-core": "^6.3.17",
-    "babel-loader": "^6.2.0",
-    "babel-polyfill": "^6.5.0",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
-    "babel-runtime": "^6.3.13",
-    "css-loader": "^0.23.0",
-    "eslint": "^1.10.3",
-    "eslint-config-airbnb": "^2.0.0",
-    "eslint-plugin-react": "^3.11.3",
-    "extract-text-webpack-plugin": "^0.9.1",
-    "file-loader": "^0.8.5",
     "isomorphic-fetch": "^2.2.0",
-    "jsdom": "^6.5.1",
-    "moment": "^2.11.2",
-    "node-sass": "^3.4.2",
     "react": "^0.14.3",
     "react-dom": "^0.14.3",
     "react-redux": "^4.0.0",
@@ -36,7 +21,22 @@
     "redux-devtools": "^3.0.0",
     "redux-devtools-dock-monitor": "^1.0.1",
     "redux-devtools-log-monitor": "^1.0.1",
-    "redux-thunk": "^1.0.0",
+    "redux-thunk": "^1.0.0"
+  },
+  "devDependencies": {
+    "babel-core": "^6.3.17",
+    "babel-loader": "^6.2.0",
+    "babel-plugin-transform-object-rest-spread": "^6.23.0",
+    "babel-polyfill": "^6.23.0",
+    "babel-preset-env": "^1.1.10",
+    "babel-preset-react": "^6.23.0",
+    "css-loader": "^0.23.0",
+    "eslint": "^1.10.3",
+    "eslint-config-airbnb": "^2.0.0",
+    "eslint-plugin-react": "^3.11.3",
+    "extract-text-webpack-plugin": "^0.9.1",
+    "file-loader": "^0.8.5",
+    "node-sass": "^3.4.2",
     "sass-lint": "^1.4.0",
     "sass-loader": "^3.1.2",
     "style-loader": "^0.13.0",
