+++ src/constants/config.js
@@ -1,5 +1,9 @@
+const socrataId = '427a-3cn5';
+const ckanId = '4582bec6-2b4f-4f9e-bc55-cbaa73117f4c';
+
 const baseConfig = {
-  SocrataURL: 'https://data.cityofboston.gov/resource/427a-3cn5'
+  SocrataURL: `https://data.cityofboston.gov/resource/${socrataId}`,
+  ckanURL: `https://data.boston.gov/api/action/datastore_search?resource_id=${ckanId}&filters=`
 };
 
 const configSwitch = {};
