+++ src/index.js
@@ -1,5 +1,3 @@
-import 'babel-polyfill';
-
 import config from './constants/config';
 import React from 'react';
 import ReactDOM from 'react-dom';
